pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pso"
version = "0.1.0"
description = "Parallel PSO benchmark - Programación Paralela"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=9.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
